Small changes to time

+++ src/project/interfaces/Barrel_C_I.java
@@ -13,5 +13,5 @@ public interface Barrel_C_I extends Remote
     public boolean isAlive() throws RemoteException;
     public void setName(String newName) throws RemoteException;
     public String getName() throws RemoteException;
-    public long getAvgExeTime() throws RemoteException;
+    public double getAvgExeTime() throws RemoteException;
 }
